#4 Merge ex-k8s test with unit test
With the refactoring all tests became unit test
